2_implementation/backend/ai-analysis-service/requirements.txt
# AI Analysis Service Dependencies
# Core Framework
fastapi==0.116.1
uvicorn[standard]==0.35.0
pydantic==2.11.7
pydantic-settings==2.10.1

# AI/ML Libraries
crewai==0.148.0
langchain==0.3.26
langchain-google-genai==2.1.8
google-generativeai==0.8.5
openai==1.3.7

# Vector Database
pymilvus==2.5.14
sentence-transformers==5.0.0

# Database
sqlalchemy==2.0.41
psycopg2-binary==2.9.10
pymongo==4.13.2
redis==6.2.0

# HTTP Client
httpx==0.28.1
aiohttp==3.12.14

# Async Support
celery==5.5.3
kombu==5.5.3

# Utilities
python-dotenv==1.1.1
python-multipart==0.0.6
python-jose[cryptography]==3.5.0
passlib[bcrypt]==1.7.4

# Data Processing
numpy==2.3.1
pandas==2.3.1
scikit-learn==1.7.1

# Testing
pytest==8.4.1
pytest-asyncio==1.1.0
pytest-cov==4.1.0
httpx==0.28.1

# Development
black==23.11.0
ruff==0.1.6
mypy==1.7.1
coverage==7.3.2

# Monitoring
structlog==23.2.0
prometheus-client==0.19.0

# Other Tools
click==8.1.7
rich==13.7.0 